+++ code/FinanceManager.Components/Features/FinancialAccounts/Components/AccountDetails.razor.cs
@@ -1,3 +1,9 @@
+using Microsoft.AspNetCore.Components;
+
 namespace FinanceManager.Components.Features.FinancialAccounts.Components;
 
-public partial class AccountDetails;
+public partial class AccountDetails
+{
+    [Parameter, SupplyParameterFromQuery(Name = "entryId")]
+    public int? EntryId { get; set; }
+}